Move DerivativesDataSink to interfaces.
